refactor internal components to -private
