+++ addon/src/components/ember-json-viewer/-private/entry-viewer.js
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
-import { isPrimitive } from '../utils/value-types';
+import { isPrimitive } from '../../../utils/value-types';
 
 export default class EntryViewer extends Component {
   @tracked
