+++ addon/src/components/ember-json-viewer/-private/simple-value.js
